+++ NEWS.md
@@ -1,5 +1,8 @@
 # surveydown (development version)
 
+- `sd_redirect()` to create redirection to external links, either by providing a button, a countdown, or both.
+- Now `sd_next()` and `sd_redirect()` both support the "Enter" key for a better user experience.
+
 # surveydown 0.2.1
 
 - In `sd_database()`, now `db_name` is changed to `dbname`, and `table_name` is changed to `table`, for consistency with Supabase, and simplicity of parameter names.